+++ lib/before-upload.interface.js
