+++ lib/before-upload.interface.metadata.json
@@ -1,0 +1,12 @@
+[
+  {
+    "__symbolic": "module",
+    "version": 3,
+    "metadata": { "UploadMetadata": { "__symbolic": "interface" } }
+  },
+  {
+    "__symbolic": "module",
+    "version": 1,
+    "metadata": { "UploadMetadata": { "__symbolic": "interface" } }
+  }
+]
