+++ lib/before-upload.interface.ts
