+++ lib/image-service.js
