+++ lib/image-upload.component.d.ts
