+++ lib/image-upload.component.metadata.json
