--- lib/image-upload/before-upload.interface.metadata.json
@@ -1,17 +1,0 @@
-[{
-  "__symbolic": "module",
-  "version": 3,
-  "metadata": {
-    "UploadMetadata": {
-      "__symbolic": "interface"
-    }
-  }
-}, {
-  "__symbolic": "module",
-  "version": 1,
-  "metadata": {
-    "UploadMetadata": {
-      "__symbolic": "interface"
-    }
-  }
-}]
