+++ lib/image.service.d.ts
