+++ lib/image.service.metadata.json
