+++ lib/style.d.ts
