+++ lib/style.js
